% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AbundanceTest}
\alias{AbundanceTest}
\title{Differential Cellular Abundance Test}
\usage{
AbundanceTest(cellmeta, celltype.col, sample.col, group.col)
}
\arguments{
\item{cellmeta}{A data frame containing metadata information about cells.}

\item{celltype.col}{The name of the column in \code{cellmeta} that contains cell type information.}

\item{sample.col}{The name of the column in \code{cellmeta} that contains sample identifiers.}

\item{group.col}{The name of the column in \code{cellmeta} that specifies the grouping variable.}
}
\value{
A data frame containing results of Wilcoxon rank sum tests for each cell type.
}
\description{
Computes a Wilcoxon rank sum test for each cell type between two groups.
}
\details{
This function takes in a data frame \code{cellmeta}, a character string \code{celltype.col},
a character string \code{sample.col}, and a character string \code{group.col} as input. It calculates
fold change, mean percentage, p-value for each cell type using Wilcoxon rank sum test.
}
\examples{
library(dplyr)
data(iris)

# Add some metadata to iris data frame
iris$Sample_ID <- rep(paste0("sample", 1:10), each = 15)
iris$Group <- rep(c("A","B"), each = 75)

# Run the AbundanceTest function
AbundanceTest(cellmeta = iris,
celltype.col = "Species",
sample.col = "Sample_ID",
group.col = "Group")

}
