% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{AddProjQual}
\alias{AddProjQual}
\title{Add projection quality scores for the query cells.}
\usage{
AddProjQual(object, k = 20, repeats = 10000)
}
\arguments{
\item{object}{A \code{CellProject} object}

\item{k}{K nearest neighbors used. Default: 20}

\item{repeats}{The number of sample times for generate background projection qualities.}
}
\value{
A \code{CellProject} object
}
\description{
Add projection quality scores for the query cells.
}
\references{
\url{https://www.nature.com/articles/s41467-021-25957-x}
}
