% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{BoxPlot}
\alias{BoxPlot}
\title{BoxPlot}
\usage{
BoxPlot(
  cellmeta,
  sample.col,
  celltype.col,
  group.col,
  legend.ncol = NULL,
  celltypes.show = NULL
)
}
\arguments{
\item{cellmeta}{A data frame containing information about the cells.}

\item{sample.col}{The column name for the sample identifier in \code{cellmeta}.}

\item{celltype.col}{The column name for the cell type in \code{cellmeta}.}

\item{group.col}{The column name for the grouping variable in \code{cellmeta}.}

\item{legend.ncol}{The number of columns for the legend (optional).}

\item{celltypes.show}{A character vector specifying which cell types to show (optional).}
}
\value{
A ggplot object.
}
\description{
Create a box plot with jittered data points and significance testing.
}
\concept{plot}
