% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CellProject-class}
\alias{CellProject-class}
\title{The CellProject class}
\description{
The main class used by Monocle to hold single cell expression data. CellDataSet extends the basic Bioconductor ExpressionSet class.
}
\details{
This class is initialized from a matrix of expression values Methods that operate on CellDataSet objects constitute the basic Monocle workflow.
}
\section{Slots}{

\describe{
\item{\code{embeddings}}{Matrix containing the predicted embeddings computed by ProjectNewdata.}

\item{\code{refined.embeddings}}{Matrix containing the refined predicted embeddings computed by RefineProjection}

\item{\code{data}}{data.frame containing the source data for cell projection.}

\item{\code{cellmeta}}{data.frame containing the data generated during the cell projection analysis.}

\item{\code{neighbors}}{\code{Neighbor} class containing the nearest neighbors for calculating the projection metrics.}
}}

\concept{class}
