% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cells.Neighbor}
\alias{cells.Neighbor}
\title{Query the nearest cells for each query cell}
\usage{
cells.Neighbor(object)
}
\arguments{
\item{object}{Neighbor object.}
}
\value{
The k nearest cell names for each query cell.
}
\description{
Query the nearest cells for each query cell
}
