% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{ComputeModuleScore}
\alias{ComputeModuleScore}
\alias{ComputeModuleScore.default}
\alias{ComputeModuleScore.Seurat}
\title{Calculate gene module scores}
\usage{
ComputeModuleScore(x, ...)

\method{ComputeModuleScore}{default}(
  x,
  gene.sets,
  min.size = 20,
  batch.size = 500,
  cores = 1,
  ...
)

\method{ComputeModuleScore}{Seurat}(
  x,
  gene.sets,
  min.size = 20,
  batch.size = 500,
  cores = 1,
  assay = Seurat::DefaultAssay(x),
  ...
)
}
\arguments{
\item{x}{gene expression matrix, rows are genes, columns are cells. Can be any format, UMI, CPM, TPM, etc.}

\item{...}{Arguments passed to other methods.}

\item{gene.sets}{a list of gene sets in data.frame or named list.}

\item{min.size}{The minimal genes of the gene sets. The size of gene sets less than this value were ignored.  Default: 20}

\item{batch.size}{The number of cells were calculated for each batch. Default: 500}

\item{cores}{number of threads for parallel computing. Default: 1}

\item{assay}{Name of the seurat object assay.}
}
\value{
A signature score matrix or Seurat object.
}
\description{
Calculate gene module scores
}
