% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateModelObject}
\alias{CreateModelObject}
\title{save models}
\usage{
CreateModelObject(model.list, gene.sets, bg.genes, ref.cellmeta)
}
\arguments{
\item{model.list}{a list of models returned from FitEnsemblSVM() or FitEnsemblMultiClassif().}

\item{gene.sets}{a list stores gene sets for calculating gene-set score matrix.}

\item{bg.genes}{a vector stores bg genes for calculating gene-set score matrix.}

\item{ref.cellmeta}{a data.frame stores cell types or other labels of reference atlas.}
}
\value{
a model class
}
\description{
save models
}
