% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{CreateReference}
\alias{CreateReference}
\title{Create reference models}
\usage{
CreateReference(
  umap.model,
  gene.sets,
  meta.data,
  bg.genes = NULL,
  gss.method = NULL,
  colors = NULL,
  text.pos = NULL
)
}
\arguments{
\item{umap.model}{a list of models returned from FitEnsemblSVM() or FitEnsemblMultiClassif().}

\item{gene.sets}{a list or data.frame stores gene sets for calculating gene-set scores.}

\item{meta.data}{a data.frame contains meta data for reference cells, including the
reference UMAP embeddings and cell labels.}

\item{bg.genes}{a vector stores background genes for calculating gene-set score matrix.}

\item{gss.method}{method for calculating gene-set scores, one of 'AUCell' and 'UCell'.}

\item{colors}{a named colors vector for visualizing the reference atlas. Deafult: NULL}

\item{text.pos}{a data.frame contains some annotated text on reference plot.}
}
\value{
a reference.model class
}
\description{
Create reference models
}
\concept{reference_mapping}
