% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{.hist2d}
\alias{.hist2d}
\title{Sum the densities of each point over the bins}
\usage{
.hist2d(x, bins, mesh.density = 1)
}
\arguments{
\item{x}{matrix, n.col=2}

\item{bins}{matrix, n.col=2}

\item{mesh.density}{length(mesh.density) equals nrow(x). Default: 1}
}
\value{
data.frame of the mesh points. (X,Y) position, bin.ID, mesh.density, bin.density.
}
\description{
Sum the densities of each point over the bins
}
\keyword{internal}
