% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{EstimateExpressionShift}
\alias{EstimateExpressionShift}
\title{Overall expression shift of the provided gene expression matrix.}
\usage{
EstimateExpressionShift(
  counts,
  cell.types,
  groups,
  n.bins = 20,
  min.cells = 10,
  n.samples = 100,
  n.cells = min.cells,
  n.cells.pairs = 45,
  cores = 1
)
}
\arguments{
\item{counts}{gene expression matrix}

\item{cell.types}{A named vector contains cell types for each cell in 'counts'}

\item{groups}{A named vector contains groups for each cell in 'counts'. Only 2 levels are surpported now.}

\item{n.bins}{number of bins for calculation NMI. Default: 20}

\item{min.cells}{minimal number of cells for each group. Default: 10}

\item{n.samples}{number of samplings for calculating expression shift of each cell type. Default: 100}

\item{n.cells}{number of cells sampled for calculation NMI in each group for each cell type. Default: min.cells}

\item{n.cells.pairs}{number of cell pairs sampled for calculation NMI on 'n.cells' sampled cells. Default: C(10,2)=45}

\item{cores}{the number of threads used. Default: 1}
}
\value{
A data.frame contains expression change between groups
}
\description{
Overall expression shift of the provided gene expression matrix.
}
