% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{EstimateKnnDensity}
\alias{EstimateKnnDensity}
\title{Estimate KNN density on 2d embedding.}
\usage{
EstimateKnnDensity(emb.mat, k = 10, n.bins = 50, n.mesh = 2, do.plot = TRUE)
}
\arguments{
\item{emb.mat}{a embedding matrix, rows are cells, columns are dimensions.}

\item{k}{Number of neighbors. Default: 10}

\item{n.bins}{number of bins for density resampling. Default: 50}

\item{n.mesh}{in each bin, density will be calculated around (n.mesh^2) points. Default: 3}

\item{do.plot}{whether plot results. Default: TRUE.}
}
\value{
a GridDensity object
}
\description{
Estimate KNN density on 2d embedding.
}
