% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{FitEnsemblMultiClassif}
\alias{FitEnsemblMultiClassif}
\title{Training an ensemble classifier for cell type prediction.}
\usage{
FitEnsemblMultiClassif(
  feature.mat,
  cell.types,
  do.norm = NULL,
  mlr3.model = "classif.svm",
  batch.size = 5000,
  n.models = 100,
  balance.cell.type = TRUE,
  cores = -1
)
}
\arguments{
\item{feature.mat}{A signature score matrix, rows are cells, columns are features.}

\item{cell.types}{A data.frame recording cell types at different granularity.}

\item{do.norm}{Whether normalize the feature matrix. L1, L2, NULL. Default: NULL.}

\item{mlr3.model}{classif.svm, classif.xgboost, classif.randomForest.  Default: classif.svm}

\item{batch.size}{The number of cells for each model. Default: 5000}

\item{n.models}{The number of SVM model. Default: 100}

\item{balance.cell.type}{A boolen determines whether performing balance sampling. Default: True}

\item{cores}{The number of CPU for training. Default: -1, use all available threads.}
}
\value{
a list of trained learners.
}
\description{
Training an ensemble classifier for cell type prediction.
}
