% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\docType{class}
\name{GridDensity-class}
\alias{GridDensity-class}
\alias{GridDensity}
\title{The GridDensity class}
\description{
The GridDensity class is an intermediate data class that storing the density of
cells in each bin (also each mesh points) and the cell assignments of the nearest
mesh point for performing the downstream analysis - production of pseudo mini bulk
RNA-seq counts and grid differential expression analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{a matrix records embedding matrix.}

\item{\code{mesh.points}}{a table records coordinates (primary key, <X,Y>), knn.density,
bin coordinates, bin.ID, and bin.density of each mesh point in data.frame.}

\item{\code{cell2mp}}{a named vector, records cell assignments to the nearest mesh point.}
}}

