% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{GroupPreferencePlot}
\alias{GroupPreferencePlot}
\title{Create Group Preference Plot}
\usage{
GroupPreferencePlot(
  cellmeta,
  group.by,
  preference.on,
  palette = "Blues",
  column_names_rot = 0,
  ...
)
}
\arguments{
\item{cellmeta}{A data frame.}

\item{group.by}{A character string indicating the name of the column in `cellmeta` used for grouping.}

\item{preference.on}{A character string indicating the name of the column in `cellmeta` used for calculating O/E ratio.}

\item{palette}{A color palette to use in the plot. Default is "Blues".}

\item{column_names_rot}{degree (0-360) to rotate x-axis labels. Default is 0.}

\item{...}{additional arguments passed to the 'Heatmap' function from ComplexHeatmap package.}
}
\value{
A heatmap that displays the O/E ratio of a two-column table grouped by a selected column.
}
\description{
A function to create a heatmap that displays the O/E ratio of a two-column table grouped by a selected column.
}
\examples{
data(mtcars)
mtcars$cyl <- factor(mtcars$cyl)
mtcars$carb <- factor(mtcars$carb)
GroupPreferencePlot(mtcars, "cyl", "carb")

}
