% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{KnnLabelTransfer}
\alias{KnnLabelTransfer}
\title{Label transfer via KNN}
\usage{
KnnLabelTransfer(query.emb, ref.emb, ref.labels, k = 100)
}
\arguments{
\item{query.emb}{The embedding matrix of query data.}

\item{ref.emb}{The embedding matrix of reference atlas.}

\item{ref.labels}{A named vector <cell name, labels>.}

\item{k}{K nearest neighbors used. Default: 100}
}
\value{
A data.frame contains prediction results.
}
\description{
Label transfer via KNN
}
\concept{label_transfer}
