% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{LabelTransfer}
\alias{LabelTransfer}
\title{A wrapper for transferring reference cell labels to the query via KNN method.}
\usage{
LabelTransfer(
  seu.q,
  reference,
  reduction.q = "ref.umap",
  ref.emb.col = paste0("UMAP_", 1:2),
  ref.label.col = "cluster.name",
  k = 10
)
}
\arguments{
\item{seu.q}{A query Seurat object.}

\item{reference}{Reference model.}

\item{reduction.q}{Reduction name storing projected reference embeddings. Default is 'ref.umap'.}

\item{ref.emb.col}{The column names storing the reference embeddings in reference model.}

\item{ref.label.col}{The column name storing the cell labels to be transferred.}

\item{k}{The K param for KNN model. Default is 10.}
}
\value{
A Seurat object. The predicted cell type is store in 'knn.pred.celltype' column.
}
\description{
A wrapper for transferring reference cell labels to the query via KNN method.
}
\concept{label_transfer}
