% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{MajorityVote}
\alias{MajorityVote}
\title{Majority voting to harmonise cell labels on given over clusters}
\usage{
MajorityVote(
  feature.mat = NULL,
  over.clusters = NULL,
  cell.types,
  k = 20,
  min.prop = 0
)
}
\arguments{
\item{feature.mat}{A signature score matrix for over clustering, rows are cells, columns are features.}

\item{over.clusters}{A named vector contains clusters for majority vote. Default: NULL}

\item{cell.types}{A data.frame recording cell types at different granularity.}

\item{k}{k of kmeans for over cluster. Default: 20}

\item{min.prop}{The minimum proportion of cells required to support naming of
the subcluster by a cell type. Default: 0}
}
\value{
A data.frame contains consensus predicted cell types.
}
\description{
Majority voting to harmonise cell labels on given over clusters
}
\concept{label_transfer}
