% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{MapQuery}
\alias{MapQuery}
\title{A wrapper for mapping query data onto reference.}
\usage{
MapQuery(
  seu.q,
  reference,
  assay.q = "RNA",
  add.map.qual = FALSE,
  ncores = 1,
  split.by = NULL
)
}
\arguments{
\item{seu.q}{A query Seurat object.}

\item{reference}{Reference model.}

\item{assay.q}{The assay used for reference mapping. Default: RNA}

\item{add.map.qual}{Whether add mapping quality metric (mean kNN distance). Default: FALSE}

\item{ncores}{Number of threads for calculation. Default: 1}

\item{split.by}{Attribute for splitting query Seurat object, projecting the query object by a list of subsetted object. 
Default: NULL}
}
\value{
A Seurat object. The projected reference embeddings were saved in a
dimension reduction object named 'ref.umap'. The gene set score were saved in
a new assay named 'SignatureScore'.
}
\description{
A wrapper for mapping query data onto reference.
}
\concept{reference_mapping}
