% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcFindAllMarkers}
\alias{mcFindAllMarkers}
\title{multi-core FindAllMarkers()}
\usage{
mcFindAllMarkers(seu, do.flatten = TRUE, only.pos = TRUE, n.cores = 10, ...)
}
\arguments{
\item{seu}{Seurat object}

\item{do.flatten}{Flattern the marker list into data.frame. Default: TRUE}

\item{only.pos}{Only positive markers. Default: TRUE}

\item{n.cores}{The number of threads. Default: 10}

\item{...}{Arguments passed to Seurat::FindMarkers()}
}
\value{
data.frame or list contains putative markers.
}
\description{
multi-core FindAllMarkers()
}
\references{
https://gist.github.com/diazdc/1735102c243cd16acb1b1f3fd09a26e1
}
