% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{merge.CellProject}
\alias{merge.CellProject}
\title{Merge CellProject Objects}
\usage{
\method{merge}{CellProject}(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{another object}

\item{...}{arguments to be passed to or from methods.}
}
\value{
merged object
}
\description{
Merge CellProject Objects
}
