% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{MergeCells}
\alias{MergeCells}
\title{Merge cells around mesh points in each bin into pseudo-bulk gene expression matrix.}
\usage{
MergeCells(dge, gd, by = "mesh.points", cores = -1)
}
\arguments{
\item{dge}{Matrix or sparse matrix for gene expression matrix rows are genes, columns are cells.}

\item{gd}{A GridDensity object.}

\item{by}{Merge raw counts by "bins" or "mesh.points". Default: "mesh.points"}

\item{cores}{Number of threads, -1 means all available threads. Default: -1}
}
\value{
A gene expression matrix, rows are genes, columns are merged cells.
}
\description{
Merge cells around mesh points in each bin into pseudo-bulk gene expression matrix.
}
