% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{newCellProject}
\alias{newCellProject}
\title{Creates a new CellProject object.}
\usage{
newCellProject(embeddings, data, cellmeta = NULL)
}
\arguments{
\item{embeddings}{matrix containing predicted embeddings}

\item{data}{matrix containing source data for cell projection}

\item{cellmeta}{data frame containing cell meta data generated during cell projection}
}
\value{
a new CellProject object
}
\description{
Creates a new CellProject object.
}
