% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NNHelper}
\alias{NNHelper}
\title{Internal helper function to dispatch to various neighbor finding methods}
\usage{
NNHelper(data, query = data, k, method, ...)
}
\arguments{
\item{data}{Input data}

\item{query}{Data to query against data}

\item{k}{Number of nearest neighbors to compute}

\item{method}{"rann", "sklearn"}

\item{...}{Other args pass \code{nn2()} or \code{sknn()}.}
}
\description{
Internal helper function to dispatch to various neighbor finding methods
}
