% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PercentageStat}
\alias{PercentageStat}
\title{Calculate cell population percentage}
\usage{
PercentageStat(cellmeta, by, fill)
}
\arguments{
\item{cellmeta}{A data frame containing cell metadata.}

\item{by}{A character vector specifying the variable(s) used to group the data.}

\item{fill}{A character vector specifying the variable used to fill the data.}
}
\value{
A data frame containing the calculated percentage statistics.
}
\description{
This function calculates percentage statistics based on a given dataset and grouping variables.
}
\examples{
PercentageStat(mtcars, "cyl", "gear")

}
\concept{differential_population}
