% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{plot.GridDensity}
\alias{plot.GridDensity}
\title{Grid density plot}
\usage{
\method{plot}{GridDensity}(x, ...)
}
\arguments{
\item{x}{A \code{GridDensity} object}

\item{...}{Arguments passed to other methods.}
}
\description{
Grid density plot
}
\concept{meta_cell}
