% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotProjection}
\alias{PlotProjection}
\title{Visualizing the projected query cells on reference via a density plot.}
\usage{
PlotProjection(
  seu.q,
  reference,
  ref.color.by,
  ref.colors = NULL,
  ref.alpha = 0.5,
  ref.size = 0.5,
  ref.emb.col = paste0("UMAP_", 1:2),
  query.reduction = "ref.umap",
  query.size = 1,
  query.alpha = 1,
  split.by = NULL,
  n.row = NULL,
  legend.ncol = 1
)
}
\arguments{
\item{seu.q}{A query Seurat object containing the projected reference embeddings.}

\item{reference}{Reference model.}

\item{ref.color.by}{Name of column in reference data to be used for coloring points.}

\item{ref.colors}{Vector of colors to be used for reference data points. Default: NULL}

\item{ref.alpha}{Alpha value for reference data points. Default: 0.5}

\item{ref.size}{Size of reference data points. Default: 0.5}

\item{ref.emb.col}{The column names storing the reference embeddings in reference model.}

\item{query.reduction}{Reduction name to use for query data. Default: 'ref.umap'}

\item{query.size}{Size of query data points. Default: 1}

\item{query.alpha}{Alpha value for query data points. Default: 1}

\item{split.by}{Column name to split query data by. Default: NULL}

\item{n.row}{Number of rows for facet_wrap when splitting by column. Default: NULL}

\item{legend.ncol}{Number of columns for legend. Default: 1}
}
\value{
A ggplot object.
}
\description{
Visualizing the projected query cells on reference via a density plot.
}
\concept{plot}
