% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{PredictNewdata}
\alias{PredictNewdata}
\title{Predict cell types}
\usage{
PredictNewdata(feature.mat, model, do.norm = NULL, cores = -1)
}
\arguments{
\item{feature.mat}{A data.frame containing signature scores, rows are cells, columns are features.}

\item{model}{The trained learners returned from `FitEnsemblMultiClassif()`.}

\item{do.norm}{Whether normalize the feature matrix. L1, L2, NULL. Default: NULL.}

\item{cores}{number of threads for prediction, -1 means all available threads. Default: -1}
}
\value{
A data.frame contains predicted results.
}
\description{
Predict cell types
}
