% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{ProjectNewdata}
\alias{ProjectNewdata}
\title{Predicting the embedding coordinates via the trained SVM model.}
\usage{
ProjectNewdata(
  feature.mat,
  model,
  do.norm = NULL,
  int.fun = stats::median,
  cores = -1
)
}
\arguments{
\item{feature.mat}{A data.frame containing signature scores, rows are cells, columns are features.}

\item{model}{The trained learners returned from `FitEnsembleSVM()`.}

\item{do.norm}{Whether normalize the feature matrix. L1, L2, NULL. Default: NULL.}

\item{int.fun}{The function for integration of predicted embedding by different learners:
mean, median, or any self defined function given a vector and returns a value. Default: median.}

\item{cores}{number of threads for prediction, -1 means all available threads. Default: -1}
}
\value{
\code{CellProject} object containing source data and predicted embedding matrix.
}
\description{
Predicting the embedding coordinates via the trained SVM model.
}
\concept{reference_mapping}
