% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{RefineProjection}
\alias{RefineProjection}
\title{Refine low confidence mapping coords.}
\usage{
RefineProjection(object, p.val.cutoff = 1, p.adj.cutoff = 0.05, k = 10)
}
\arguments{
\item{object}{A \code{CellProject} object}

\item{p.val.cutoff}{Interpolation for the coordinates of cells with p.value
above this cutoff. Default: 1}

\item{p.adj.cutoff}{Interpolation for the coordinates of cells with p.value
above this cutoff. Default: 0.05}

\item{k}{K nearest neighbors for interpolation. Default: 10}
}
\value{
A \code{CellProject} object
}
\description{
Refine low confidence mapping coords.
}
