% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ribo.genes}
\alias{ribo.genes}
\title{A data set containing all human ribosomal genes}
\format{
A vector with 184 elements:
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/GOCC_RIBOSOME.html}
}
\usage{
ribo.genes
}
\description{
A data set containing all human ribosomal genes
}
\concept{data}
\keyword{datasets}
