% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sknn}
\alias{sknn}
\title{Run sklearn.neighbors.NearestNeighbors}
\usage{
sknn(data, query = data, k = 5L, ...)
}
\arguments{
\item{data}{Input data}

\item{query}{Data to query against data}

\item{k}{Number of nearest neighbors to compute}

\item{...}{other parameters pass to \code{sklearn.neighbors.NearestNeighbors()}.}
}
\description{
Run sklearn.neighbors.NearestNeighbors
}
