% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{SplitCellProject}
\alias{SplitCellProject}
\title{Split CellProject Object}
\usage{
SplitCellProject(x, split.by)
}
\arguments{
\item{x}{CellProject object}

\item{split.by}{Attribute for splitting}
}
\value{
A list of CellProject object
}
\description{
Split CellProject Object
}
