% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cells.R
\name{subset.GridDensity}
\alias{subset.GridDensity}
\title{subset a GridDensity object.}
\usage{
\method{subset}{GridDensity}(
  x,
  bin.density.threshold = NULL,
  mesh.density.threshold = -1,
  n.cells.threshold = 10,
  n.mesh.threshold = -1,
  ...
)
}
\arguments{
\item{x, }{GridDensity object}

\item{bin.density.threshold}{Only bins with density greater than this value are retained. Default: mean-3*sd}

\item{mesh.density.threshold}{Only mesh points with density greater than this value are retained.
We do not filter on mesh point directly. Default: -1}

\item{n.cells.threshold}{Only mesh points with assigned cells no less than this value are retained. Default: 10}

\item{n.mesh.threshold}{Only bins with number of mesh pints no less than this value are retained. Default: -1}

\item{...}{Arguments passed to other methods.
Default: max number of mesh points per bin minus 1.}
}
\description{
subset a GridDensity object.
}
