% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{subset2}
\alias{subset2}
\title{Subset CellProject Object}
\usage{
subset2(x, cells)
}
\arguments{
\item{x}{A CellProject object}

\item{cells}{cells for subsetting}
}
\value{
A subsetted CellProject object
}
\description{
Subset CellProject Object
}
\concept{object}
