% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{VarDecompose}
\alias{VarDecompose}
\title{Find Differential Expressed Genes via Variance Deomposition using Mixed Linear Model}
\usage{
VarDecompose(data, meta.data, vd.vars, genes = "all", cores = -1)
}
\arguments{
\item{data}{A data.frame or matrix of gene expression matrix}

\item{meta.data}{A data.frame contains cell meta data}

\item{vd.vars}{variable in meta.data for variance decomposition.}

\item{genes}{Genes for variance decomposition. Default: "all"}

\item{cores}{The number of threads for calculation, -1 means all available
threads. Default: -1}
}
\value{
A data.frame contains the results of variance decomposition.
}
\description{
Find Differential Expressed Genes via Variance Deomposition using Mixed Linear Model
}
