% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Plot the Volcano Plot}
\usage{
VolcanoPlot(
  test.df,
  p.value.cutoff = 0.001,
  xlab = NULL,
  ylab = NULL,
  colors = NULL
)
}
\arguments{
\item{test.df}{A data frame from `AbundanceTest()`.}

\item{p.value.cutoff}{The p-value cutoff for showing the cell type labels.
Default: 0.001}

\item{xlab}{The label for the x-axis. Default: NULL}

\item{ylab}{The label for the y-axis. Default: NULL}

\item{colors}{A vector of colors to be used in the plot.}
}
\value{
A ggplot object.
}
\description{
Volcano plot for visualizing the cellular abundance difference. This function
accepts the outputs of `AbundanceTest()`.
}
