% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{FitEnsembleSVM}
\alias{FitEnsembleSVM}
\title{Training an ensemble SVM model for embedding coordinates regression.}
\usage{
FitEnsembleSVM(
  feature.mat,
  emb.mat,
  do.norm = NULL,
  batch.size = 5000,
  n.models = 100,
  cores = -1
)
}
\arguments{
\item{feature.mat}{A signature score matrix, rows are cells, columns are features.}

\item{emb.mat}{The embedding matrix, rows are cells, columns are dimensions.}

\item{do.norm}{Whether normalize the feature matrix. L1, L2, NULL. Default: NULL.}

\item{batch.size}{The number of cells for each model. Default: 5000}

\item{n.models}{The number of SVM model. Default: 100}

\item{cores}{The number of CPU for training. Default: -1, use all available threads.}
}
\value{
a list of trained learners.
}
\description{
Training an ensemble SVM model for embedding coordinates regression.
}
