% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{KnnLabelTransfer}
\alias{KnnLabelTransfer}
\title{Label transfer via KNN}
\usage{
KnnLabelTransfer(pred.emb, ref.emb, ref.celltype, k = 100)
}
\arguments{
\item{pred.emb}{The predicted embedding matrix of query data.}

\item{ref.emb}{The embedding matrix of reference atlas.}

\item{ref.celltype}{A named vector <name, value> = <cellname, celltype>.}

\item{k}{K nearest neighbors used. Default: 100}
}
\value{
A data.frame contains prediction results.
}
\description{
Label transfer via KNN
}
